+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## NEXT
 
+Improvements:
 * [MacOS] Added the corners and size information to barcode results.
 * [MacOS] Added support for `analyzeImage`.
 * [MacOS] Added a Privacy Manifest.
 * [web] Added the size information to barcode results.
 * Added support for barcode formats to image analysis.
+* Updated the scanner to report any scanning errors that were encountered during processing.
+
+Bugs fixed:
+* Fixed a bug that would cause the scanner to emit an error when it was already started. Now it ignores any calls to start while it is starting.
+* [Android] Fixed a bug that prevented `useNewCameraSelector` from being used correctly. (thanks @bswhite1!)
 
 ## 5.2.3