+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Improvements:
 * [web] Added the size information to barcode results.
 * Added support for barcode formats to image analysis.
 * Updated the scanner to report any scanning errors that were encountered during processing.
+* Introduced a new getter `hasCameraPermission` for the `MobileScannerState`.
+* Fixed a bug in the lifecycle handling sample. Now instead of checking `isInitialized`,
+the sample recommends using `hasCameraPermission`, which also guards against camera permission errors.
 
 Bugs fixed:
 * Fixed a bug that would cause the scanner to emit an error when it was already started. Now it ignores any calls to start while it is starting.